--- src/Types/Type.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Picqer\Barcode\Types;
-
-abstract class Type
-{
-    abstract public function getBarcodeData(string $code): array;
-}
