+++ src/Types/TypeInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Picqer\Barcode\Types;
+
+interface TypeInterface
+{
+    public function getBarcodeData(string $code): array;
+}
